+++ src/data/db/migrationRunner.ts
@@ -25,6 +25,7 @@ import * as migration022 from './migrations/022_pre_reading_reflection_run';
 import * as migration023 from './migrations/023_book_capsule_run';
 import * as migration024 from './migrations/024_dnf_reflection_run';
 import * as migration025 from './migrations/025_rating_run';
+import * as migration026 from './migrations/026_hot_query_indexes';
 
 const log = createLogger('db/migrations');
 
@@ -70,6 +71,7 @@ const migrations: Migration[] = [
   { version: migration023.version, up: migration023.up },
   { version: migration024.version, up: migration024.up, manualTransaction: migration024.manualTransaction },
   { version: migration025.version, up: migration025.up, manualTransaction: migration025.manualTransaction },
+  { version: migration026.version, up: migration026.up },
 ];
 
 export const LATEST_SCHEMA_VERSION = migrations[migrations.length - 1]?.version ?? 0;